refactor auth commands
