+++ packages/wrangler/src/__tests__/sentry.test.ts
@@ -259,11 +259,22 @@ describe("sentry", () => {
 				              Object {
 				                "colno": 0,
 				                "context_line": "",
-				                "filename": "/wrangler/packages/wrangler/src/index.ts",
+				                "filename": "/wrangler/packages/wrangler/src/core/wrap-command.ts",
 				                "function": "",
 				                "in_app": false,
 				                "lineno": 0,
-				                "module": "index.ts",
+				                "module": "wrap-command.ts",
+				                "post_context": Array [],
+				                "pre_context": Array [],
+				              },
+				              Object {
+				                "colno": 0,
+				                "context_line": "",
+				                "filename": "/wrangler/packages/wrangler/src/auth/commands.ts",
+				                "function": "",
+				                "in_app": false,
+				                "lineno": 0,
+				                "module": "commands.ts",
 				                "post_context": Array [],
 				                "pre_context": Array [],
 				              },
