+++ packages/wrangler/src/commands.ts
@@ -1,0 +1,2 @@
+import "./auth/commands";
+import "./kv/index";
